small refactor to sdk for better readability - typescript and organization
