+++ src/SpotIMAPI.ts
@@ -1,93 +1,16 @@
-import { NativeEventEmitter, NativeModules, PixelRatio, Platform, UIManager, findNodeHandle, requireNativeComponent } from 'react-native';
-
-import PropTypes from 'prop-types';
-import React from 'react';
+import { NativeModules, Platform } from 'react-native';
+import { SpotIMEventEmitter } from './SpotIMEventEmitter';
 
 const ISIOS = Platform.OS === "ios";
 
-export const SpotIMModule = NativeModules.SpotIM;
-const SpotIMEvents = NativeModules.SpotIMEvents;
-export const SpotIMEventEmitter = new NativeEventEmitter(SpotIMEvents);
-
-export class SpotIM extends React.Component {
-
-  nativeComponentRef;
-  lastHeightUpdate = 0;
-  heightUpdateTimestamp;
-
-  constructor(props) {
-    super(props);
-    this._onChange = this._onChange.bind(this);
-  }
-
-  componentDidMount() {
-    if (ISIOS) {
-      SpotIMEventEmitter.addListener('viewHeightDidChange', (event) => {
-        this.setState({ height: event['newHeight'] });
-      });
-    } else {
-      this.loadConversationAndroid()
-    }
-  }
-
-  loadConversationAndroid() {
-    this.setState({ height: 0 });
-    this.lastHeightUpdate = 0;
-    this.dispatchCreateCommandAndroid();
-    this.setState({ loading: true });
-    this.heightUpdateTimestamp = Date.now() / 1000;
-    setTimeout(() => {
-      this.setState({ loading: false });
-    }, 2500);
-  }
-
-  dispatchCreateCommandAndroid() {
-    const androidViewId = findNodeHandle(this.nativeComponentRef);
-    UIManager.dispatchViewManagerCommand(
-      androidViewId,
-      UIManager.SpotIM.Commands.create.toString(),
-      [androidViewId]
-    );
-  }
-  _onChange(event) {
-    const newHeight = event.nativeEvent.height / PixelRatio.get();
-    const heightDiff = Math.abs(newHeight - this.lastHeightUpdate);
-    if ((!this.lastHeightUpdate || heightDiff != 0)) {
-      this.lastHeightUpdate = newHeight;
-      if (!this.state.height ||
-        this.state.height < newHeight ||
-        Date.now() / 1000 - this.heightUpdateTimestamp > 5) {
-        this.setState({ height: newHeight });
-      }
-      this.heightUpdateTimestamp = Date.now() / 1000;
-    }
-
-  }
-  componentWillUnmount() {
-    if (ISIOS) {
-      SpotIMEventEmitter.removeAllListeners("viewHeightDidChange")
-    }
-  }
-
-  componentDidUpdate(prevProps) {
-    if (!ISIOS && (prevProps.postId != this.props.postId)) {
-      this.loadConversationAndroid()
-    }
-  }
-
-  render() {
-    return <RNSpotIM
-      {...this.props}
-      onChange={this._onChange}
-      ref={(nativeRef) => this.nativeComponentRef = nativeRef}
-      style={{ alignSelf: 'stretch', height: this.state && this.state.height ? Number(this.state.height) : 0, opacity: this.state && this.state.loading ? 0 : 1 }} />;
-  }
-}
+const SpotIMModule = NativeModules.SpotIM;
 
 export class SpotIMAPI {
-  static init = (spotId) => {
+
+  static init = (spotId: string) => {
     SpotIMModule.initWithSpotId(spotId)
   }
+
   static startSSO = () => {
     return new Promise((resolve, reject) => {
       const successSubscription = SpotIMEventEmitter.addListener('startSSOSuccess', (response) => {
@@ -105,7 +28,8 @@ export class SpotIMAPI {
       SpotIMModule.startSSO();
     })
   }
-  static completeSSO = (str) => {
+
+  static completeSSO = (str: string) => {
     return new Promise((resolve, reject) => {
       const successSubscription = SpotIMEventEmitter.addListener('completeSSOSuccess', (response) => {
         successSubscription.remove();
@@ -122,7 +46,8 @@ export class SpotIMAPI {
       SpotIMModule.completeSSO(str);
     })
   }
-  static sso = (jwt) => {
+
+  static sso = (jwt: string) => {
     return new Promise((resolve, reject) => {
       const successSubscription = SpotIMEventEmitter.addListener('ssoSuccess', (response) => {
         successSubscription.remove();
@@ -139,6 +64,7 @@ export class SpotIMAPI {
       SpotIMModule.ssoWithJwtSecret(jwt);
     })
   }
+
   static getUserLoginStatus = () => {
     return new Promise((resolve, reject) => {
       const successSubscription = SpotIMEventEmitter.addListener('getUserLoginStatusSuccess', (response) => {
@@ -156,6 +82,7 @@ export class SpotIMAPI {
       SpotIMModule.getUserLoginStatus();
     })
   }
+
   static logout = () => {
     return new Promise((resolve, reject) => {
       const successSubscription = SpotIMEventEmitter.addListener('logoutSuccess', (response) => {
@@ -173,24 +100,14 @@ export class SpotIMAPI {
       SpotIMModule.logout();
     })
   }
+
   static showFullConversation = () => {
     SpotIMModule.showFullConversation()
   }
-  static setIOSDarkMode = (isOn) => {
+
+  static setIOSDarkMode = (isOn: boolean) => {
     if (ISIOS) {
       SpotIMModule.setIsDarkMode(isOn)
     }
   }
 }
-
-SpotIM.propTypes = {
-  postId: PropTypes.string,
-  url: PropTypes.string,
-  title: PropTypes.string,
-  subtitle: PropTypes.string,
-  thumbnailUrl: PropTypes.string
-};
-
-var RNSpotIM = requireNativeComponent('SpotIM', SpotIM, {
-  nativeOnly: { onChange: true }
-});
