+++ src/SpotIMEventEmitter.ts
@@ -1,0 +1,4 @@
+import { NativeModules, NativeEventEmitter } from 'react-native';
+
+const SpotIMEvents = NativeModules.SpotIMEvents;
+export const SpotIMEventEmitter = new NativeEventEmitter(SpotIMEvents);
