+++ src/index.ts
@@ -1,0 +1,3 @@
+export { SpotIMAPI } from './SpotIMAPI';
+export { SpotIM } from './SpotIM';
+export { SpotIMEventEmitter } from './SpotIMEventEmitter';
